Move order related fixtures from global fixture file
