+++ saleor/order/tests/fixtures/__init__.py
@@ -1,0 +1,2 @@
+from .order import *  # noqa: F403
+from .order_line import *  # noqa: F403
